+++ yolo_bit/include/wifi_mqtt.h
@@ -1,0 +1,8 @@
+#ifndef WIFI_MQTT_H
+#define WIFI_MQTT_H
+
+#include "global.h"
+
+void wifi_mqtt_task(void *pvParameters);
+
+#endif // WIFI_MQTT_H