+++ yolo_bit/docs/CHANGELOG.md
@@ -8,7 +8,8 @@
 - **Added Temp/Humi Task:** Implemented in `temp_humi_monitor.cpp/h`.
 - **Added Secured Door Task:** Scaffolding implemented in `secured_door.cpp/h`.
 - **Added AppContext:** Utilizing Semaphores for IPC in `global.h` & `config.h`.
-- **Change include** paths in `global.h` from:
+- **Added WiFi & MQTT Task:** Integrated `PubSubClient` and `<WiFi.h>` in `wifi_mqtt.cpp/h` to publish real-time telemetry and subscribe to Web Service commands. Updated dependent files (`global.h`, `main.cpp`, `config.h`) to share global sensor data using memory mutexes.
+- **Changed include** paths in `global.h` from:
 
 ```cpp
 //this cause a not-found error
@@ -19,5 +20,5 @@
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
 #include "freertos/semphr.h"
-//this fix everything
+//this fixes everything
 ```